requirements.txt
asgiref==3.5.0
contourpy==1.2.1
cycler==0.12.1
Django==4.0.1
django-cors-headers==3.11.0
django-pwa==1.0.10
djangorestframework==3.15.1
fonttools==4.51.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.2.2
scipy==1.13.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.5.0
tzdata==2024.1
